State for determining which sections to render is now elevated to CardDialog from CardDialogContentBody.
